relocating booking/computer endpoint
