+++ controllers/booking/parser.py
@@ -2,7 +2,5 @@ from flask_restful import reqparse
 
 def booking_parser():
     parser = reqparse.RequestParser()
-    parser.add_argument('id', type=str, required=True, help="Este campo no puede estar en blanco")
-    parser.add_argument('student_email', type=str, required=True, help="Este campo no puede estar en blanco")
-    parser.add_argument('computer', type=str, required=True, help="Este campo no puede estar en blanco")
+    
     return parser
